+++ packages/grammarly-api/CHANGELOG.md
@@ -1,4 +1,4 @@
-# unofficial-grammarly-api
+# unofficial-grammarly-api-2
 
 ## 0.2.0