+++ packages/grammarly-language-server/CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Patch Changes
 
 - Updated dependencies [1ed857d]
-  - unofficial-grammarly-api@0.2.0
+  - unofficial-grammarly-api-2@0.2.0
 
 ## 0.1.0
 
@@ -20,4 +20,4 @@
 ### Patch Changes
 
 - Updated dependencies [undefined]
-  - unofficial-grammarly-api@0.1.0
+  - unofficial-grammarly-api-2@0.1.0